Replacing click selecting entirely with drag selecting since it now behaves the same
In this change, I fixed the selection box sizing and registered a
bounds checking event handler for each unit so that box selection may
be conducted. This way, selection is always performed with a list of
units to be selected, even if the list only contains 1 unit. The list
is then passed to the game controller, where it may be filtered by
priority in a future change. Above all else, it’s cool done this way.